Split transfer-related code into downloads.py and uploads.py
Makes things easier to find and maintain.
