+++ pynicotine/gtkgui/dialogs/fileproperties.py
@@ -160,7 +160,7 @@ class FileProperties(Dialog):
 
         properties = self.properties[self.current_index]
 
-        core.transfers.get_file(
+        core.downloads.get_file(
             properties["user"], properties["file_path"], size=properties["size"],
             file_attributes=properties.get("file_attributes")
         )
